requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
clip==1.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
flash_attn==2.6.3
fonttools==4.55.0
frozenlist==1.5.0
ftfy==6.3.1
huggingface-hub==0.26.2
matplotlib==3.9.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.24.3
open_clip_torch==2.29.0
openai==1.54.3
packaging==24.1
pandas==2.2.3
platformdirs==4.3.6
qwen-vl-utils==0.0.8
requests==2.32.3
safetensors==0.4.5
six==1.16.0
smmap==5.0.1
sympy==1.13.1
termcolor==2.5.0
tokenizers==0.20.3
torch==torch2.3.0
torchvision==0.18.0
tqdm==4.66.6
transformers==4.46.0